update to chromium v60.0.3112.20, fixes build problem.

+++ app/src/main/java/org/chromium/chrome/browser/ntp/RecentTabsManager.java
@@ -5,30 +5,28 @@
 package org.chromium.chrome.browser.ntp;
 
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.graphics.Bitmap;
 
-import org.chromium.base.ObserverList;
+import org.chromium.base.ContextUtils;
 import org.chromium.base.ThreadUtils;
-import org.chromium.chrome.browser.UrlConstants;
+import org.chromium.base.VisibleForTesting;
+import org.chromium.base.metrics.RecordUserAction;
 import org.chromium.chrome.browser.favicon.FaviconHelper;
 import org.chromium.chrome.browser.favicon.FaviconHelper.FaviconImageCallback;
-import org.chromium.chrome.browser.firstrun.ProfileDataCache;
+import org.chromium.chrome.browser.history.HistoryManagerUtils;
 import org.chromium.chrome.browser.invalidation.InvalidationController;
+import org.chromium.chrome.browser.metrics.StartupMetrics;
 import org.chromium.chrome.browser.ntp.ForeignSessionHelper.ForeignSession;
 import org.chromium.chrome.browser.ntp.ForeignSessionHelper.ForeignSessionCallback;
 import org.chromium.chrome.browser.ntp.ForeignSessionHelper.ForeignSessionTab;
-import org.chromium.chrome.browser.ntp.RecentTabsPromoView.SyncPromoModel;
-import org.chromium.chrome.browser.ntp.RecentlyClosedBridge.RecentlyClosedCallback;
-import org.chromium.chrome.browser.ntp.RecentlyClosedBridge.RecentlyClosedTab;
 import org.chromium.chrome.browser.profiles.Profile;
 import org.chromium.chrome.browser.signin.SigninManager;
 import org.chromium.chrome.browser.signin.SigninManager.SignInStateObserver;
-import org.chromium.chrome.browser.sync.SyncController;
 import org.chromium.chrome.browser.tab.Tab;
-import org.chromium.content_public.browser.LoadUrlParams;
-import org.chromium.sync.AndroidSyncSettings;
-import org.chromium.sync.AndroidSyncSettings.AndroidSyncSettingsObserver;
-import org.chromium.sync.signin.ChromeSigninController;
+import org.chromium.components.signin.ChromeSigninController;
+import org.chromium.components.sync.AndroidSyncSettings;
+import org.chromium.components.sync.AndroidSyncSettings.AndroidSyncSettingsObserver;
 
 import java.util.Collections;
 import java.util.List;
@@ -36,8 +34,7 @@ import java.util.List;
 /**
  * Provides the domain logic and data for RecentTabsPage and RecentTabsRowAdapter.
  */
-public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInStateObserver,
-        SyncPromoModel {
+public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInStateObserver {
 
     /**
      * Implement this to receive updates when the page contents change.
@@ -50,22 +47,24 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
     }
 
     private static final int RECENTLY_CLOSED_MAX_TAB_COUNT = 5;
+    private static final String PREF_SIGNIN_PROMO_DECLINED =
+            "recent_tabs_signin_promo_declined";
+
+    private static RecentlyClosedTabManager sRecentlyClosedTabManagerForTests;
 
     private final Profile mProfile;
     private final Tab mTab;
     private final Context mContext;
-    private final ObserverList<AndroidSyncSettingsObserver> mObservers =
-            new ObserverList<AndroidSyncSettingsObserver>();
 
     private FaviconHelper mFaviconHelper;
     private ForeignSessionHelper mForeignSessionHelper;
     private List<ForeignSession> mForeignSessions;
     private List<RecentlyClosedTab> mRecentlyClosedTabs;
-    private NewTabPagePrefs mNewTabPagePrefs;
-    private RecentlyClosedBridge mRecentlyClosedBridge;
+    private RecentTabsPagePrefs mPrefs;
+    private RecentlyClosedTabManager mRecentlyClosedTabManager;
     private SigninManager mSignInManager;
     private UpdatedCallback mUpdatedCallback;
-    private ProfileDataCache mProfileDataCache;
+    private boolean mIsDestroyed;
 
     /**
      * Create an RecentTabsManager to be used with RecentTabsPage and RecentTabsRowAdapter.
@@ -77,12 +76,22 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
     public RecentTabsManager(Tab tab, Profile profile, Context context) {
         mProfile = profile;
         mTab = tab;
-        mForeignSessionHelper = buildForeignSessionHelper(mProfile);
-        mNewTabPagePrefs = buildNewTabPagePrefs(mProfile);
-        mFaviconHelper = buildFaviconHelper();
-        mRecentlyClosedBridge = buildRecentlyClosedBridge(mProfile);
+        mForeignSessionHelper = new ForeignSessionHelper(profile);
+        mPrefs = new RecentTabsPagePrefs(profile);
+        mFaviconHelper = new FaviconHelper();
+        mRecentlyClosedTabManager = sRecentlyClosedTabManagerForTests != null
+                ? sRecentlyClosedTabManagerForTests
+                : new RecentlyClosedBridge(profile);
         mSignInManager = SigninManager.get(context);
         mContext = context;
+
+        mRecentlyClosedTabManager.setTabsUpdatedRunnable(new Runnable() {
+            @Override
+            public void run() {
+                updateRecentlyClosedTabs();
+                postUpdate();
+            }
+        });
 
         updateRecentlyClosedTabs();
         registerForForeignSessionUpdates();
@@ -97,6 +106,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * Should be called when this object is no longer needed. Performs necessary listener tear down.
      */
     public void destroy() {
+        mIsDestroyed = true;
         AndroidSyncSettings.unregisterObserver(mContext, this);
 
         mSignInManager.removeSignInStateObserver(this);
@@ -105,47 +115,18 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
         mFaviconHelper.destroy();
         mFaviconHelper = null;
 
-        mRecentlyClosedBridge.destroy();
-        mRecentlyClosedBridge = null;
+        mRecentlyClosedTabManager.destroy();
+        mRecentlyClosedTabManager = null;
 
         mForeignSessionHelper.destroy();
         mForeignSessionHelper = null;
 
         mUpdatedCallback = null;
 
-        mNewTabPagePrefs.destroy();
-        mNewTabPagePrefs = null;
-
-        if (mProfileDataCache != null) {
-            mProfileDataCache.destroy();
-            mProfileDataCache = null;
-        }
+        mPrefs.destroy();
+        mPrefs = null;
 
         InvalidationController.get(mContext).onRecentTabsPageClosed();
-    }
-
-    private static ForeignSessionHelper buildForeignSessionHelper(Profile profile) {
-        return new ForeignSessionHelper(profile);
-    }
-
-    private static NewTabPagePrefs buildNewTabPagePrefs(Profile profile) {
-        return new NewTabPagePrefs(profile);
-    }
-
-    private static FaviconHelper buildFaviconHelper() {
-        return new FaviconHelper();
-    }
-
-    private RecentlyClosedBridge buildRecentlyClosedBridge(Profile profile) {
-        RecentlyClosedBridge bridge = new RecentlyClosedBridge(profile);
-        bridge.setRecentlyClosedCallback(new RecentlyClosedCallback() {
-            @Override
-            public void onUpdated() {
-                updateRecentlyClosedTabs();
-                postUpdate();
-            }
-        });
-        return bridge;
     }
 
     private void registerForForeignSessionUpdates() {
@@ -163,12 +144,9 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
         mSignInManager.addSignInStateObserver(this);
     }
 
-    protected void updateCurrentlyOpenTabs() {
-    }
-
     private void updateRecentlyClosedTabs() {
-        mRecentlyClosedTabs = mRecentlyClosedBridge.getRecentlyClosedTabs(
-                RECENTLY_CLOSED_MAX_TAB_COUNT);
+        mRecentlyClosedTabs =
+                mRecentlyClosedTabManager.getRecentlyClosedTabs(RECENTLY_CLOSED_MAX_TAB_COUNT);
     }
 
     private void updateForeignSessions() {
@@ -176,13 +154,6 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
         if (mForeignSessions == null) {
             mForeignSessions = Collections.emptyList();
         }
-    }
-
-    /**
-     * @return Most up-to-date list of currently open tabs.
-     */
-    public List<CurrentlyOpenTab> getCurrentlyOpenTabs() {
-        return null;
     }
 
     /**
@@ -208,7 +179,8 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      */
     public void openForeignSessionTab(ForeignSession session, ForeignSessionTab tab,
             int windowDisposition) {
-        NewTabPageUma.recordAction(NewTabPageUma.ACTION_OPENED_FOREIGN_SESSION);
+        if (mIsDestroyed) return;
+        RecordUserAction.record("MobileRecentTabManagerTabFromOtherDeviceOpened");
         mForeignSessionHelper.openForeignSessionTab(mTab, session, tab, windowDisposition);
     }
 
@@ -220,15 +192,18 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      *         be restored into the current tab or a new tab.
      */
     public void openRecentlyClosedTab(RecentlyClosedTab tab, int windowDisposition) {
-        NewTabPageUma.recordAction(NewTabPageUma.ACTION_OPENED_RECENTLY_CLOSED_ENTRY);
-        mRecentlyClosedBridge.openRecentlyClosedTab(mTab, tab, windowDisposition);
+        if (mIsDestroyed) return;
+        RecordUserAction.record("MobileRecentTabManagerRecentTabOpened");
+        mRecentlyClosedTabManager.openRecentlyClosedTab(mTab, tab, windowDisposition);
     }
 
     /**
      * Opens the history page.
      */
     public void openHistoryPage() {
-        mTab.loadUrl(new LoadUrlParams(UrlConstants.HISTORY_URL));
+        if (mIsDestroyed) return;
+        HistoryManagerUtils.showHistoryManager(mTab.getActivity(), mTab);
+        StartupMetrics.getInstance().recordOpenedHistory();
     }
 
     /**
@@ -266,55 +241,14 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
     }
 
     /**
-     * Sets the persistent expanded/collapsed state of the currently open tabs list.
-     *
-     * @param isCollapsed Whether the currently open tabs list is collapsed.
-     */
-    public void setCurrentlyOpenTabsCollapsed(boolean isCollapsed) {
-        mNewTabPagePrefs.setCurrentlyOpenTabsCollapsed(isCollapsed);
-    }
-
-    /**
-     * Determine the expanded/collapsed state of the currently open tabs list.
-     *
-     * @return Whether the currently open tabs list is collapsed.
-     */
-    public boolean isCurrentlyOpenTabsCollapsed() {
-        return mNewTabPagePrefs.getCurrentlyOpenTabsCollapsed();
-    }
-
-    /**
-     * Sets the state for showing all tabs in the currently open tabs list. This is intended to
-     * be overridden in extending classes and set to true when the user clicks the "More" button
-     * at the end of the list.
-     * @param showingAll Whether the currently open tabs list should start to show all.
-     */
-    public void setCurrentlyOpenTabsShowAll(boolean showingAll) {
-    }
-
-    /**
-     * @return Whether the currently open tabs group shows all tabs. If it is not, only a limited
-     * number of tabs is shown with a "More" button at the end of the list to show all.
-     */
-    public boolean isCurrentlyOpenTabsShowingAll() {
-        return false;
-    }
-
-    /**
-     * Closes the specified currently open tab.
-     * @param tab Information about the tab that should be closed.
-     */
-    public void closeTab(CurrentlyOpenTab tab) {
-    }
-
-    /**
      * Sets the persistent expanded/collapsed state of a foreign session list.
      *
      * @param session foreign session to collapsed.
      * @param isCollapsed Whether the session is collapsed or expanded.
      */
     public void setForeignSessionCollapsed(ForeignSession session, boolean isCollapsed) {
-        mNewTabPagePrefs.setForeignSessionCollapsed(session, isCollapsed);
+        if (mIsDestroyed) return;
+        mPrefs.setForeignSessionCollapsed(session, isCollapsed);
     }
 
     /**
@@ -325,7 +259,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @return Whether the session is collapsed.
      */
     public boolean getForeignSessionCollapsed(ForeignSession session) {
-        return mNewTabPagePrefs.getForeignSessionCollapsed(session);
+        return mPrefs.getForeignSessionCollapsed(session);
     }
 
     /**
@@ -334,7 +268,8 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the recently closed tabs list is collapsed.
      */
     public void setRecentlyClosedTabsCollapsed(boolean isCollapsed) {
-        mNewTabPagePrefs.setRecentlyClosedTabsCollapsed(isCollapsed);
+        if (mIsDestroyed) return;
+        mPrefs.setRecentlyClosedTabsCollapsed(isCollapsed);
     }
 
     /**
@@ -343,10 +278,10 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @return Whether the recently closed tabs list is collapsed.
      */
     public boolean isRecentlyClosedTabsCollapsed() {
-        return mNewTabPagePrefs.getRecentlyClosedTabsCollapsed();
+        return mPrefs.getRecentlyClosedTabsCollapsed();
     }
 
-   /**
+    /**
      * Remove Foreign session to display. Note that it might reappear during the next sync if the
      * session is not orphaned.
      *
@@ -354,6 +289,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param session Session to be deleted.
      */
     public void deleteForeignSession(ForeignSession session) {
+        if (mIsDestroyed) return;
         mForeignSessionHelper.deleteForeignSession(session);
     }
 
@@ -361,7 +297,8 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * Clears the list of recently closed tabs.
      */
     public void clearRecentlyClosedTabs() {
-        mRecentlyClosedBridge.clearRecentlyClosedTabs();
+        if (mIsDestroyed) return;
+        mRecentlyClosedTabManager.clearRecentlyClosedTabs();
     }
 
     /**
@@ -370,11 +307,27 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @return Whether sync promo should be displayed.
      */
     public boolean shouldDisplaySyncPromo() {
-        if (SigninManager.get(mContext).isSigninDisabledByPolicy()) {
+        SigninManager signinManager = SigninManager.get(mContext);
+        if (signinManager.isSigninDisabledByPolicy() || !signinManager.isSigninSupported()) {
+            return false;
+        }
+
+        if (ContextUtils.getAppSharedPreferences().getBoolean(
+                PREF_SIGNIN_PROMO_DECLINED, false)) {
             return false;
         }
 
         return !AndroidSyncSettings.isSyncEnabled(mContext) || mForeignSessions.isEmpty();
+    }
+
+    /**
+     * Save that user tapped "No" button on the signin promo.
+     */
+    public void setSigninPromoDeclined() {
+        SharedPreferences.Editor sharedPreferencesEditor =
+                ContextUtils.getAppSharedPreferences().edit();
+        sharedPreferencesEditor.putBoolean(PREF_SIGNIN_PROMO_DECLINED, true);
+        sharedPreferencesEditor.apply();
     }
 
     /**
@@ -383,7 +336,8 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the sync promo is collapsed.
      */
     public void setSyncPromoCollapsed(boolean isCollapsed) {
-        mNewTabPagePrefs.setSyncPromoCollapsed(isCollapsed);
+        if (mIsDestroyed) return;
+        mPrefs.setSyncPromoCollapsed(isCollapsed);
     }
 
     /**
@@ -392,10 +346,10 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @return Whether the sync promo is collapsed.
      */
     public boolean isSyncPromoCollapsed() {
-        return mNewTabPagePrefs.getSyncPromoCollapsed();
+        return mPrefs.getSyncPromoCollapsed();
     }
 
-    protected void postUpdate() {
+    private void postUpdate() {
         if (mUpdatedCallback != null) {
             mUpdatedCallback.onUpdated();
         }
@@ -418,46 +372,19 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
         ThreadUtils.runOnUiThread(new Runnable() {
             @Override
             public void run() {
+                if (mIsDestroyed) return;
                 updateForeignSessions();
                 postUpdate();
-                for (AndroidSyncSettingsObserver observer : mObservers) {
-                    observer.androidSyncSettingsChanged();
-                }
             }
         });
     }
 
-    // SyncPromoModel
-    @Override
-    public boolean isSyncEnabled() {
-        return AndroidSyncSettings.isSyncEnabled(mContext);
-    }
-
-    @Override
     public boolean isSignedIn() {
-        return ChromeSigninController.get(mContext).isSignedIn();
+        return ChromeSigninController.get().isSignedIn();
     }
 
-    @Override
-    public void enableSync() {
-        SyncController.get(mContext).start();
-    }
-
-    @Override
-    public void registerForSyncUpdates(AndroidSyncSettingsObserver changeListener) {
-        mObservers.addObserver(changeListener);
-    }
-
-    @Override
-    public void unregisterForSyncUpdates(AndroidSyncSettingsObserver changeListener) {
-        mObservers.removeObserver(changeListener);
-    }
-
-    @Override
-    public ProfileDataCache getProfileDataCache() {
-        if (mProfileDataCache == null) {
-            mProfileDataCache = new ProfileDataCache(mContext, Profile.getLastUsedProfile());
-        }
-        return mProfileDataCache;
+    @VisibleForTesting
+    public static void setRecentlyClosedTabManagerForTests(RecentlyClosedTabManager manager) {
+        sRecentlyClosedTabManagerForTests = manager;
     }
 }
